// lib/types/sanity.types.ts
export type SocialMediaAccount = {
  _type: 'socialMediaAccount';
  label: 'Facebook' | 'Twitter' | 'Instagram' | 'LinkedIn' | 'X';
  href: string;
};
// ...
export type NavigationItem = {
  _type: 'navigationItem';
  label?: string;
  href: string;
  subnav?: Array<
    {
      _key: string;
    } & NavigationItem
  >;
};

export type Navigation = {
  _id: string;
  _type: 'navigation';
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  title?: string;
  findYourSpace?: Array<
    {
      _key: string;
    } & NavigationItem
  >;
  featuredLocations?: Array<
    {
      _key: string;
    } & NavigationItem
  >;
  whyUs?: Array<
    {
      _key: string;
    } & NavigationItem
  >;
  existingCustomers?: Array<
    {
      _key: string;
    } & NavigationItem
  >;
  aboutUs?: Array<
    {
      _key: string;
    } & NavigationItem
  >;
};
// ...
export type Globals = {
  _id: string;
  _type: 'globals';
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  siteName?: string;
  footerLinks?: Array<
    {
      _key: string;
    } & FooterLink
  >;
  socialMedia?: Array<
    {
      _key: string;
    } & SocialMediaAccount
  >;
};
// ...
export type FooterLink = {
  _type: 'footerLink';
  label: string;
  href: string;
};